乗車Phase2: RidingTestHelper を整理しテストを DI 解決へ
RidingTestHelper を SeatedTrainCarGuid と RegisterSeatedCarOnNewTrain のみに削減。
CreateEnvironment が DI コンテナを生成しているのに new TrainUnitDatastore() 等で
別インスタンスを生成していたのはテスト規約違反だったため、各テストは
ServiceProvider.GetService で IPlayerRidingDatastore / RidableResolver を解決し、
RegisterSeatedCarOnNewTrain は DI 上の TrainUnitDatastore へ車両を登録する。
接続状態の差し替えが要る OnRidableRemoved テストのみ、checker を注入した
PlayerRidingDatastore を明示構築する（DI 既定では切断を再現できないため）。
